Operadores aritiméticos

+++ src/academy/devdojo/maratonajava/introducao/Aula04Operadores.java
@@ -5,7 +5,7 @@ public class Aula04Operadores {
         // + - / *
         int numero01 = 10;
         int numero02 = 20;
-        double resultado = numero01 / numero02;
+        int resultado = numero01 / numero02;
         System.out.println(resultado);
     }
 }
